requirements.txt
python-dotenv~=1.0.1
langchain~=0.2.16
langchain-openai~=0.1.23
langchain-core~=0.2.38
requests~=2.32.3
langchain-community~=0.2.16
pydantic~=2.9.1